# Speed up slowest test