--- a/test/features/goals_test.exs
+++ b/test/features/goals_test.exs
@@ -57,9 +57,9 @@ defmodule Operately.Features.GoalTest do
     |> Steps.close_goal(params)
     |> Steps.assert_goal_closed(params)
     |> Steps.assert_goal_is_not_editable()
-    # |> Steps.assert_goal_closed_email_sent()
-    # |> Steps.assert_goal_closed_feed_posted()
-    # |> Steps.assert_goal_closed_notification_sent()
+    |> Steps.assert_goal_closed_email_sent()
+    |> Steps.assert_goal_closed_feed_posted()
+    |> Steps.assert_goal_closed_notification_sent()
   end
 
   @tag login_as: :champion