I've noticed that some UI tests are excessively slow. For example this one:

![Screenshot 2024-06-25 at 18 23 52](https://github.com/operately/operately/assets/1779493/d092aa17-b8eb-4b0e-808e-1d2f5aeda251)

I went down the rabbithole to figure out the reason.

Turns out, this test was heavily using `Wallaby.Browser.refute_has`, which in turn is calling https://github.com/elixir-wallaby/wallaby/blob/main/lib/wallaby/browser.ex#L1496. I've put an IO.inspect into that function and noticed that it is reaching out to the running browser hundreds of times. 

Basically, it was trying to find the element for 3 seconds, and if it didn't found it, it declared that refute_has(text) has passed. In the original test, we were running 4 times in a row refute_has, amount to 12 seconds of needless waiting.

---

the new implementation of the `UI.refute_text` now simply gets all the available text on the screen, and tests if the desired text is visible. This reduced the execution time to below 0.1s, making the original test 5 seconds in total, instead of 19 seconds.